+++ Promises/Internal/MergeInternal.cs
@@ -121,7 +121,7 @@ namespace Proto.Promises
             partial class MergePromise<T> : IInvokable
             {
                 // These are used to avoid rounding errors when normalizing the progress.
-                private float _expected;
+                private uint _expected;
                 private UnsignedFixed32 _currentAmount;
                 private bool _invokingProgress;
                 private bool _suspended;
